Added decryption error handling, fixed search bug. Added some error handling for vault encryption and decryption, fixing a bug where if the data that was being decrypted wasn't valid, namely, the length of the data not being a multiple of the AES block size, an unhandled exception would be thrown.
Also fixed bug where when accounts are being searched for, and an
account is tapped, rather than opening the entry viewer for the index
relative to the front buffer, it would open relative to the back buffer,
so the account opened would not correspond to the account tapped. Fixed
by using the dedicated itemAt for front/back buffers added in previous
commit to VaultAccountAdapter.