Enforced portrait mode temporarily, and tweaked permissions.
Due to the activity's being reset whenever the orientation changes,
deleting potentially unsaved changes, portrait mode is being temporarily
enforced to prevent the orientation from changing in the first place.

This is a temporary solution, and eventually a state saver will have to
be implemented, but until then, this should prevent data loss.

Also, permissions were tweaked in MainActivity