Fixed a very, very sneaky bug. Explicit truncate mode added.
Previously, neither "a" (append) or "t" (truncate) mode were specified
for the file descriptor, only "w" (write), however, on other versions of
Android, or perhaps other API levels, or implementations, the default
mode is not truncate, but append, leading to very sneaky data mismatches
when the new data is smaller than the data that's being overwritten.

The solution: to explicitly specify truncate mode.

Shoutout to Temina.